Move GroupHeadings.ts to src/Query/
