+++ src/Query/TaskGroups.ts
@@ -1,6 +1,6 @@
-import { GroupHeadings } from '../GroupHeadings';
 import type { Grouping } from '../Query';
 import type { Task } from '../Task';
+import { GroupHeadings } from './GroupHeadings';
 import { IntermediateTaskGroups } from './IntermediateTaskGroups';
 import { TaskGroup } from './TaskGroup';
